Handle uint64 from generate_post FFI call (#1228)
* update rust-proofs submodule

* seal output and verify_seal input now expect CommRStar

* drop the fake SectorStore arg

* proof of replication now uses 2 * 192 bytes

* drop now-unused Storage field

* use uint64 for fault indices instead of uint8

generate_post may be supplied more than 256 replica commitments, so uint8 wasn't appropriate

* to be safe, make it a little larger

* update comments

* update rust-proofs submodule

* bump seal output (proof) to 384 bytes

* seal now takes longer - bump up the timeout

+++ protocol/storage/miner.go
@@ -339,7 +339,7 @@ func (sm *Miner) getProvingPeriodStart() (*types.BlockHeight, error) {
 // generatePoSt creates the required PoSt, given a list of sector ids and
 // matching seeds. It returns the Snark Proof for the PoSt, and a list of
 // sectors that faulted, if there were any faults.
-func generatePoSt(commRs [][32]byte, seed [32]byte) ([192]byte, []uint8, error) {
+func generatePoSt(commRs [][32]byte, seed [32]byte) ([192]byte, []uint64, error) {
 	req := proofs.GeneratePoSTRequest{
 		CommRs:        commRs,
 		ChallengeSeed: seed,
